requirements.txt
requests>=2.28.0
pyyaml>=6.0
langchain-core>=0.3.0
langchain-huggingface>=0.1.0
langchain-google-genai>=2.0.0
langchain-openai>=0.3.0
langgraph>=0.1.0